requirements.txt
mss==9.0.1
pillow==10.0.1
pyautogui==0.9.54
websockets==12.0